refactor(admin): unify geo moderation surface for screenshot triage
Consolidate the four stacked status banners into a single
ModerationStatusRail (counter row + cold-start / readiness / run-state
sections sharing one health subscription).

Replace the three add-map dialogs (Manual / Wand / Research) with one
AddMapDialog that exposes the strategies as tabs; the side-panel
buttons drive the controlled strategy so each entry lands on the right
tab without remounting.

Fold the standalone Reports tab into the Moderation tab as a sub-tab,
rename the top-level entry to "Modération", and redirect the legacy
?tab=reports URL to ?tab=geo so existing bookmarks land on the right
sub-section.

Net: 18 files (~4800 LOC) → 11 files; 4 stacked banners → 1 rail; 3
modal dialogs → 1; 9 admin tabs → 8.
